requirements.txt
requests
flask
flask-session
peewee
gunicorn
six
praw
qrcode
Pillow
arrow
